command: execute系の関数を別ファイルに移動する
command.goが肥大化してきたため
